requirements.txt
Django==1.11.6
Markdown==2.6.9
Pillow==4.3.0
PyYAML==3.12
asgi-redis==1.4.3
awsebcli==3.12.0
blessed==1.14.2
boto3==1.4.7
botocore==1.7.28
cement==2.8.2
colorama==0.3.7
channels==1.1.8
django-background-tasks==1.1.13
django-filter==1.1.0
django-storages==1.6.5
djangorestframework==3.7.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
numpy==1.13.3
olefile==0.44
pathspec==0.5.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
requests==2.9.1
s3transfer==0.1.11
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
tensorflow==1.5.1
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.44.0